trying to bugfix
i have dug myself a tunnel too deep to continue.
I need to step back a bit and rethink how i'm gonna approach this.
I'll probably have to change some of the codes structure.
Some of my matrix math is not checking out. I either need to use a public lib to handle it for me or find and fix the issue, but i have no idea where to look next. I've been at this for far to long. I will do one last restructure to try and fix the problem.