smoothed cam movement and optimized input handling
there is now an input class singleton that handles all inputs.
using this input system the camera movement is now smoothed out.

The new controls to move around are WASD or arrow keys to move around the x/z plane. and QE or + and - to move on the Y axis.

This will change once rotations through the mouse are inserted. The inputs will then correlate the the cameras up/right/forward vectors instead of the coordinate system.