uvsphere normals and small shader preperations
i added the normals to the uvsphere.
I had struggles with the shaders. Unfortunately i started fresh again because i f'd up and it was to hard to find the mistake. I'm basically back at zero when it comes to lighting my scene properly with Blinn-Phong shading. I do managed to understand the concept better now and  should be able to implement it without any further holdups. I am just to tired to do it now. I hope i can manage in time for the deadline.