blur highlight renderoutput
I now have all needed components for the blur effect. The last step is to blur the bloom texture and combining it with the standard texture. Unfortunatly i need the FBOs to function so i can render into a texture before combining!!!
Let's hope i can get this right in time...